backend/requirements.txt
Django==3.1.6

django-cors-headers==3.7.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
drf-social-oauth2==1.0.9

gunicorn==20.0.4
psycopg2-binary==2.8.6
